requirements.txt
Flask==3.0.0
pytelegrambotapi==4.17.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
flask_sqlalchemy==3.1.1